requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.4.9
openai==1.12.0
google-generativeai==0.3.2
aiohttp==3.9.1
cachetools==5.3.2
tenacity==8.2.3
redis==5.0.1
python-dotenv==1.0.0